fix: types for Context and plugins, remove blank lines
