+++ src/plugins/OauthPlugin/Oauth/Oauth.js
@@ -9,7 +9,7 @@ const { ERRORS } = ConnectError;
 
 export default class Oauth {
   /**
-   * @param {OauthOptions} params Params for constructor
+   * @param {OauthOptionsWithStrategy} params Params for constructor
    */
   constructor({ clientId, scopes, oauthServer, frameStrategy, oauthStrategy }) {
     this.clientId = clientId;
@@ -23,6 +23,9 @@ export default class Oauth {
     this.setScopes(scopes);
   }
 
+  /**
+   * @returns {string}
+   */
   get storeId() {
     return `endpass-oauth:${this.clientId}`;
   }
@@ -31,7 +34,8 @@ export default class Oauth {
    * Initiate token
    * @deprecated
    * @param {object} params Parameters object
-   * @param {string[]} params.scopes - Array of authorization scopes
+   * @param {string[]} [params.scopes] - Array of authorization scopes
+   * @returns {Promise<void>}
    */
   async loginWithOauth(params) {
     this.setScopes(params.scopes);
@@ -39,8 +43,8 @@ export default class Oauth {
   }
 
   /**
-   *
    * @param {string[]=} scopes
+   * @returns {void}
    */
   setScopes(scopes) {
     if (!scopes) {
@@ -50,6 +54,9 @@ export default class Oauth {
     this.checkScopes();
   }
 
+  /**
+   * @returns {void}
+   */
   checkScopes() {
     const tokenObject = this.getTokenObjectFromStore();
     const now = new Date().getTime();
@@ -108,6 +115,9 @@ export default class Oauth {
     return tokenObject;
   }
 
+  /**
+   * @returns {void}
+   */
   logout() {
     LocalStorage.remove(this.storeId);
   }
@@ -154,8 +164,8 @@ export default class Oauth {
 
   /**
    * Makes api request with authorization token
-   * @template {import('axios').AxiosRequestConfig} T
-   * @param {{scopes?: string[]} & T} params
+   * @param {OauthRequestOptions} params
+   * @returns {import('axios').AxiosPromise<any>}
    */
   request({ scopes, ...options }) {
     this.setScopes(scopes);
