+++ types/Request.d.ts
@@ -1,3 +1,5 @@
+/// <reference path="constants.d.ts" />
+
 declare type OriginReq = {
   method: string,
   answer: Function,
@@ -20,3 +22,6 @@ declare type EventResult = {
   error: Error,
   code: ConnectErrors[keyof ConnectErrors]
 }
+
+declare type RequestMethods = PluginMethods & MessengerMethods & Web3Methods & InpageEvents;
+declare type RequestMethodsValues = RequestMethods[keyof RequestMethods];
